+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix: DRAIN has a standard range in 6e.
 - Fix: Partial STR dice is now added to damage. [#1193](https://github.com/dmdorman/hero6e-foundryvtt/issues/1193)
 - Fixed TRANSFORM to import properly, allowing for proper attack & damage rolls and END cost.
+- Fix STUN for END calculation especially when starting with negative END. [#1202](https://github.com/dmdorman/hero6e-foundryvtt/issues/1202)
 
 ## Version 3.0.92 [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)